+++ external-crates/move/tooling/prettier-move/src/index.ts
@@ -34,7 +34,7 @@ export const languages: SupportLanguage[] = [
 	{
 		name: 'move',
 		extensions: ['.move'],
-		parsers: ['move-parse'],
+		parsers: ['move'],
 	},
 ];