+++ external-crates/move/tooling/prettier-move/CHANGELOG.md
@@ -4,10 +4,10 @@
 
 ### Patch Changes
 
--   e1a85c2: fixes publishing issue, compiles prepublish
+- e1a85c2: fixes publishing issue, compiles prepublish
 
 ## 0.1.0
 
 ### Minor Changes
 
--   9521492: Initial version of the prettier-plugin-move
+- 9521492: Initial version of the prettier-plugin-move